+++ Tests/FirebaseBundleTest.php
@@ -11,7 +11,7 @@ use PHPUnit\Framework\TestCase;
 class FirebaseBundleTest extends TestCase
 {
     /** @test */
-    public function it_uses_the_right_container_extension()
+    public function it_uses_the_right_container_extension(): void
     {
         $bundle = new FirebaseBundle();